// src/shared/constants/response-message.enum.ts
export enum RESPONSE_MESSAGE {
  HEALTH_CHECK_SUCCESS = "HEALTH_CHECK_SUCCESS",
  INTERNAL_ERROR = "INTERNAL_ERROR",
  ROUTE_NOT_FOUND = "ROUTE_NOT_FOUND",
  UNAUTHORIZED = "UNAUTHORIZED",
  FORBIDDEN = "FORBIDDEN",

  OTP_RESEND_SUCCESS = "OTP_RESEND_SUCCESS",
  OTP_RESEND_FAILED = "OTP_RESEND_FAILED",
  OTP_RESEND_COOLDOWN_ACTIVE = "OTP_RESEND_COOLDOWN_ACTIVE",

  LOGIN_SUCCESS = "LOGIN_SUCCESS",
  LOGIN_FAILED = "LOGIN_FAILED",
  LOGOUT_SUCCESS = "LOGOUT_SUCCESS",
  LOGOUT_FAILED = "LOGOUT_FAILED",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  USER_NOT_FOUND = "USER_NOT_FOUND",

  VALIDATION_FAILED = "VALIDATION_FAILED",
  REGISTRATION_SUCCESS = "REGISTRATION_SUCCESS",
  PHONE_ALREADY_REGISTERED = "PHONE_ALREADY_REGISTERED",
  EMAIL_ALREADY_REGISTERED = "EMAIL_ALREADY_REGISTERED",

  OTP_VERIFIED = "OTP_VERIFIED",
  OTP_EXPIRED = "OTP_EXPIRED",
  OTP_INVALID = "OTP_INVALID",

  PROFILE_FETCHED = "PROFILE_FETCHED",
  PROFILE_NOT_FOUND = "PROFILE_NOT_FOUND",

  // SUB - ORGANIZATION

  SUB_ORGANIZATION_CREATED = "SUB_ORGANIZATION_CREATED",
  SUB_ORGANIZATION_CREATION_FAILED = "SUB_ORGANIZATION_CREATION_FAILED",
  SUB_ORGANIZATIONS_FETCHED = "SUB_ORGANIZATIONS_FETCHED",
  SUB_ORGANIZATIONS_FETCH_FAILED = "SUB_ORGANIZATIONS_FETCH_FAILED",
  SUB_ORGANIZATION_UPDATED = "SUB_ORGANIZATION_UPDATED",
  SUB_ORGANIZATION_UPDATE_FAILED = "SUB_ORGANIZATION_UPDATE_FAILED",
  SUB_ORGANIZATION_NOT_FOUND = "SUB_ORGANIZATION_NOT_FOUND",
  SUB_ORGANIZATION_BLOCKED = "SUB_ORGANIZATION_BLOCKED",
  SUB_ORGANIZATION_BLOCK_FAILED = "SUB_ORGANIZATION_BLOCK_FAILED",

  PASSWORD_UPDATE_FAILED = "PASSWORD_UPDATE_FAILED",
  PASSWORD_UPDATED = "PASSWORD_UPDATED",
  CURRENT_PASSWORD_INCORRECT = "CURRENT_PASSWORD_INCORRECT",

  EVENT_DRAFT_SAVE_FAILED = "EVENT_DRAFT_SAVE_FAILED",
  EVENT_DRAFT_SAVED = "EVENT_DRAFT_SAVED",
  EVENT_NOT_READY_TO_PUBLISH = "EVENT_NOT_READY_TO_PUBLISH",
  EVENT_PUBLISHED = "EVENT_PUBLISHED",
  EVENT_FETCHED = "EVENT_FETCHED",
  EVENT_NOT_FOUND = "EVENT_NOT_FOUND",

  // ADMIN - ORGANIZATION

  ORGANIZATIONS_FETCHED = "ORGANIZATIONS_FETCHED",
  ORGANIZATIONS_FETCH_FAILED = "ORGANIZATIONS_FETCH_FAILED",
  ORGANIZATION_FETCHED = "ORGANIZATION_FETCHED",
  ORGANIZATION_FETCH_FAILED = "ORGANIZATION_FETCH_FAILED",

  ORGANIZATION_REGISTRATION_ALLOWED = "ORGANIZATION_REGISTRATION_ALLOWED",
  ORGANIZATION_REGISTRATION_ALLOW_FAILED = "ORGANIZATION_REGISTRATION_ALLOW_FAILED",
  ORGANIZATION_REGISTRATION_REJECTED = "ORGANIZATION_REGISTRATION_REJECTED",
  ORGANIZATION_REGISTRATION_REJECT_FAILED = "ORGANIZATION_REGISTRATION_REJECT_FAILED",

  ORGANIZATION_BLOCKED = "ORGANIZATION_BLOCKED",
  ORGANIZATION_BLOCK_FAILED = "ORGANIZATION_BLOCK_FAILED",
  ORGANIZATION_UNBLOCKED = "ORGANIZATION_UNBLOCKED",
  ORGANIZATION_UNBLOCK_FAILED = "ORGANIZATION_UNBLOCK_FAILED",

  PARTICIPANTS_FETCHED = "PARTICIPANTS_FETCHED",

  // ADMIN - USERS

  USERS_FETCHED = "USERS_FETCHED",
  USERS_FETCH_FAILED = "USERS_FETCH_FAILED",
  USER_BLOCKED = "USER_BLOCKED",
  USER_BLOCK_FAILED = "USER_BLOCK_FAILED",
  USER_UNBLOCKED = "USER_UNBLOCKED",
  USER_UNBLOCK_FAILED = "USER_UNBLOCK_FAILED",
}